+++ src/app/components/Experience.tsx
@@ -1,18 +1,80 @@
-import React from "react";
-import { ItemExp } from "./ItemExp";
+"use client"; // this is a client component 👈🏽
+import { motion } from "framer-motion";
 
-const Experience = () => (
-  <section className="s2">
-    <div className="main-container">
-      <h3>Experiencia Laboral</h3>
-      <p>Esta es la experiencia que adquiri en estos años 🤓</p>
-      <div className="flex gap-2 flex-wrap">
-        <ItemExp />
-        <ItemExp />
-        <ItemExp />
+import Image from "next/image";
+import "react-vertical-timeline-component/style.min.css";
+
+import {
+  VerticalTimeline,
+  VerticalTimelineElement,
+} from "react-vertical-timeline-component";
+import { experiences } from "@/store/data";
+import { textVariant } from "@/utils";
+
+const ExperienceCard = ({ experience }: any) => {
+  return (
+    <VerticalTimelineElement
+      contentStyle={{
+        background: "var(--previewShadow)",
+        color: "#fff",
+      }}
+      contentArrowStyle={{ borderRight: "7px solid  #232631" }}
+      date={experience.date}
+      iconStyle={{ background: experience.iconBg }}
+      icon={
+        <div className="flex justify-center items-center w-full h-full">
+          <Image
+            src={experience.icon}
+            alt={experience.company_name}
+            className="w-[60%] h-[60%] object-contain"
+          />
+        </div>
+      }
+    >
+      <div>
+        <h3 className="text-white text-[24px] font-bold">{experience.title}</h3>
+        <p
+          className="text-secondary text-[16px] font-semibold"
+          style={{ margin: 0 }}
+        >
+          {experience.company_name}
+        </p>
+      </div>
+
+      <ul className="mt-5 list-disc ml-5 space-y-2">
+        {experience.points.map((point: any, index: number) => (
+          <li
+            key={`experience-point-${index}`}
+            className="text-white-100 text-[14px] pl-1 tracking-wider"
+          >
+            {point}
+          </li>
+        ))}
+      </ul>
+    </VerticalTimelineElement>
+  );
+};
+
+const Experience = () => {
+  return (
+    <div className="pt-20 bg-[--mainColor]">
+      <motion.div className="text-center" variants={textVariant(1)}>
+        <span>en lo que eh estado trabajando</span>
+        <h2>Experiencia Profesional:</h2>
+      </motion.div>
+
+      <div className="mt-20 flex flex-col">
+        <VerticalTimeline>
+          {experiences.map((experience, index) => (
+            <ExperienceCard
+              key={`experience-${index}`}
+              experience={experience}
+            />
+          ))}
+        </VerticalTimeline>
       </div>
     </div>
-  </section>
-);
+  );
+};
 
 export default Experience;
