+++ src/app/globals.css
@@ -555,8 +555,8 @@ h6 {
 .about-wrapper {
   display: -ms-grid;
   display: grid;
-  -ms-grid-columns: (minmax(320px, 1fr))[auto-fit];
-  grid-template-columns: repeat(auto-fit, minmax(320px, 1fr));
+  -ms-grid-columns: (minmax(300px, 1fr))[auto-fit];
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
   padding-bottom: 100px;
   padding-top: 100px;
   gap: 100px;
